fixing issue related to Intel video driver on Windows, along with some performance analysis/experimentation/improvement

+++ src/main_es.frag
@@ -1,7 +1,7 @@
 #version 100
 
-precision highp float;
-precision highp sampler2D;
+precision mediump float;
+precision mediump sampler2D;
 
 varying vec4 frag_Color;
 varying vec2 frag_Tex;
